Do not exit when table store drops the connection
The `os.Exit(1)` doesn't allow a user to control what to do on an error, like a retry for example. Stopping the program execution is a major problem when running long running jobs like an import.

While it would be cleaner to return the actual error this change keeps the API compatibility and gives a good indication of the failure. Where for further debugging the `dumpSessions` is still available.